refactor(spec): split binary structures into modules
